Refactor code into light module and update project structure
