+++ Lightning_practice/.gitignore
@@ -172,4 +172,3 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
-**data**
